+++ apps/meteor/app/apps/server/bridges/livechat.ts
@@ -17,6 +17,7 @@ import { registerGuest } from '../../../livechat/server/lib/guests';
 import type { ILivechatMessage } from '../../../livechat/server/lib/localTypes';
 import { updateMessage, sendMessage } from '../../../livechat/server/lib/messages';
 import { createRoom } from '../../../livechat/server/lib/rooms';
+import { transfer } from '../../../livechat/server/lib/transfer';
 import { settings } from '../../../settings/server';
 
 declare module '@rocket.chat/apps/dist/converters/IAppMessagesConverter' {
@@ -277,7 +278,7 @@ export class AppLivechatBridge extends LivechatBridge {
 		}
 
 		// #TODO: #AppsEngineTypes - Remove explicit types and typecasts once the apps-engine definition/implementation mismatch is fixed.
-		return LivechatTyped.transfer(
+		return transfer(
 			(await this.orch.getConverters()?.get('rooms').convertAppRoom(currentRoom)) as IOmnichannelRoom,
 			this.orch.getConverters()?.get('visitors').convertAppVisitor(visitor),
 			{ userId, departmentId, transferredBy, transferredTo },
