+++ apps/meteor/app/livechat/server/lib/stream/agentStatus.ts
@@ -1,8 +1,8 @@
 import { Logger } from '@rocket.chat/logger';
 
 import { settings } from '../../../../settings/server';
-import { Livechat } from '../LivechatTyped';
 import { closeOpenChats } from '../closeRoom';
+import { forwardOpenChats } from '../transfer';
 
 const logger = new Logger('AgentStatusWatcher');
 
@@ -73,7 +73,7 @@ export const onlineAgents = {
 			}
 
 			if (action === 'forward') {
-				return await Livechat.forwardOpenChats(userId);
+				return await forwardOpenChats(userId);
 			}
 		} catch (e) {
 			logger.error({
